separating PDF from image controllers
