+++ site/app/controllers/ImageController.php
@@ -1,0 +1,242 @@
+<?php
+
+namespace app\controllers;
+
+use app\libraries\Core;
+use app\controllers\AbstractController;
+use app\libraries\FileUtils;
+use app\libraries\response\JsonResponse;
+use Symfony\Component\Routing\Annotation\Route;
+use app\models\User;
+
+class ImageController extends AbstractController {
+    public function __construct(Core $core) {
+        parent::__construct($core);
+    }
+
+    /**
+     * Load annotation JSON data for a given file path from the annotation directory
+     * @param string $annotation_dir The directory containing annotation files
+     * @param string $file_path The file path to generate MD5 hash for matching
+     * @param int|false|null $latest_timestamp Reference to track latest modification time (optional)
+     * @return array<mixed> Array of annotation JSONs keyed by grader_id
+     */
+    private function loadAnnotationJsons(string $annotation_dir, string $file_path, &$latest_timestamp = null): array {
+        $annotation_jsons = [];
+        $file_path_md5 = md5($file_path);
+
+        if (is_dir($annotation_dir)) {
+            $dir_iter = new \FilesystemIterator($annotation_dir);
+            foreach ($dir_iter as $annotation_file) {
+                if (explode('_', $annotation_file->getFilename())[0] === $file_path_md5) {
+                    $file_contents = file_get_contents($annotation_file->getPathname());
+                    $annotation_decoded = json_decode($file_contents, true);
+                    if ($annotation_decoded !== null) {
+                        $grader_id = $annotation_decoded["grader_id"];
+                        $annotation_jsons[$grader_id] = json_encode($annotation_decoded['annotations']);
+
+                        // Track latest timestamp if requested
+                        if ($latest_timestamp !== null && $latest_timestamp < $annotation_file->getMTime()) {
+                            $latest_timestamp = $annotation_file->getMTime();
+                        }
+                    }
+                }
+            }
+        }
+
+        return $annotation_jsons;
+    }
+
+    /**
+     * returns the path contains the anonymous id given the gradeable id if $real is false
+     * returns the path for the submitted file given the user id if $real is true
+     */
+    public function getPath(string $file_path, string $id, bool $real): string {
+        $path = "";
+        $file_path_parts = explode("/", $file_path);
+
+        if ($real && str_contains($file_path, "..")) {
+            return "INVALID FILE PATH";
+        }
+
+        for ($index = 1; $index < count($file_path_parts); $index++) {
+            if ($index === 9) {
+                if ($real) {
+                    $path .= "/" . $id;
+                }
+                else {
+                    $user_id = $file_path_parts[$index];
+                    $anon_ids = $this->core->getQueries()->getSubmitterIdFromAnonId($user_id, $id);
+                    $path .= "/" . (empty($anon_ids) ? $user_id : $anon_ids[$user_id]);
+                }
+            }
+            else {
+                $path = $path . "/" . $file_path_parts[$index];
+            }
+        }
+        return $path;
+    }
+
+    #[Route("/courses/{_semester}/{_course}/gradeable/{gradeable_id}/img")]
+    public function showStudentImg(string $gradeable_id, string $filename, string $path, string $anon_path): JsonResponse {
+        $id = $this->core->getUser()->getId();
+        $gradeable = $this->tryGetGradeable($gradeable_id);
+        if ($gradeable === false) {
+            return JsonResponse::getErrorResponse('Could not get gradeable');
+        }
+        if ($gradeable->isTeamAssignment()) {
+            $id = $this->core->getQueries()->getTeamByGradeableAndUser($gradeable_id, $id)->getId();
+        }
+        $submitter = $this->core->getQueries()->getSubmitterById($id);
+        $graded_gradeable = $this->core->getQueries()->getGradedGradeableForSubmitter($gradeable, $submitter);
+        $active_version = $graded_gradeable->getAutoGradedGradeable()->getActiveVersion();
+
+        $annotation_path = FileUtils::joinPaths($this->core->getConfig()->getCoursePath(), 'annotations', $gradeable_id, $id, $active_version);
+        $annotation_jsons = $this->loadAnnotationJsons($annotation_path, $path);
+
+        // Construct the image URL for display_file endpoint
+        $image_url = $this->core->buildCourseUrl(['display_file']) .
+            '?dir=submissions_processed' .
+            '&file=' . urlencode($filename) .
+            '&path=' . urlencode($path) .
+            '&ta_grading=true' .
+            '&view_annotated_file=true';
+
+        return JsonResponse::getSuccessResponse([
+            'image_url' => $image_url,
+            'filename' => $filename,
+            'annotations' => $annotation_jsons,
+            'gradeable_id' => $gradeable_id,
+            'user_id' => $id
+        ]);
+    }
+
+    #[Route(path: "/courses/{_semester}/{_course}/gradeable/{gradeable_id}/grading/img", methods: ["POST"])]
+    public function showGraderImageEmbedded(string $gradeable_id): void {
+        // User can be a team
+        $id = $_POST['user_id'] ?? null;
+        $filename = $_POST['filename'] ?? null;
+        $filename = html_entity_decode($filename);
+        // this file_path is assumed to be an anonymous path
+        $file_path = urldecode($_POST['file_path']);
+
+        $id = $this->core->getQueries()->getSubmitterIdFromAnonId($id, $gradeable_id);
+        $real_path = $this->getPath($file_path, $id, true);
+        if (!file_exists($real_path)) {
+            $this->core->getOutput()->renderJsonFail('The image file could not be found');
+            return;
+        }
+
+        $gradeable = $this->tryGetGradeable($gradeable_id);
+        if ($gradeable === false) {
+            $this->core->getOutput()->renderJsonFail('Could not get gradeable');
+            return;
+        }
+
+        $graded_gradeable = $this->tryGetGradedGradeable($gradeable, $id);
+        if ($graded_gradeable === false) {
+            $this->core->getOutput()->renderJsonFail('Could not get graded gradeable');
+            return;
+        }
+
+        if (!$this->core->getAccess()->canI("grading.electronic.grade", ["gradeable" => $gradeable, "graded_gradeable" => $graded_gradeable])) {
+            $this->core->getOutput()->renderJsonFail('You do not have permission to grade this student');
+            return;
+        }
+
+        // We've already verified that we can grade this assignment.  We just check to see if this a peer grader
+        // to determine if we should show a button to download the image.
+        $is_peer_grader = $this->core->getUser()->getGroup() === User::GROUP_STUDENT && $gradeable->hasPeerComponent();
+
+        $active_version = $graded_gradeable->getAutoGradedGradeable()->getActiveVersion();
+        $annotation_dir = FileUtils::joinPaths($this->core->getConfig()->getCoursePath(), 'annotations', $gradeable_id, $id, $active_version);
+        $annotation_jsons = $this->loadAnnotationJsons($annotation_dir, $file_path);
+
+        $this->core->getOutput()->renderOutput(['Image'], 'showImageEmbedded', $gradeable_id, $id, $filename, $file_path, $file_path, $file_path, $annotation_jsons, false, $is_peer_grader);
+    }
+
+    #[Route("/courses/{_semester}/{_course}/gradeable/{gradeable_id}/img/{target_dir}", methods: ["POST"])]
+    public function saveImageAnnotation(string $gradeable_id, string $target_dir): JsonResponse {
+        // Save image annotations similar to PDF annotations
+        $grader_id = $this->core->getUser()->getId();
+        $course_path = $this->core->getConfig()->getCoursePath();
+        $user_id = $_POST['user_id'] ?? null;
+        $filename = $_POST['filename'] ?? null;
+        $file_path = $_POST['file_path'] ?? null;
+        $annotations = $_POST['annotations'] ?? '[]';
+
+        if ($user_id === null || $filename === null || $file_path === null) {
+            return JsonResponse::getErrorResponse('Missing required parameters');
+        }
+
+        $gradeable = $this->tryGetGradeable($gradeable_id);
+        if ($gradeable === false) {
+            return JsonResponse::getErrorResponse('Could not get gradeable');
+        }
+
+        $graded_gradeable = $this->tryGetGradedGradeable($gradeable, $user_id);
+        if ($graded_gradeable === false) {
+            return JsonResponse::getErrorResponse('Could not get graded gradeable');
+        }
+
+        // Leaving the target_dir parameter gives us flexibility in the future, but it is currently only allowed to
+        // ever have one value ("annotations").
+        if ($target_dir !== 'annotations') {
+            return JsonResponse::getErrorResponse("Invalid target directory $target_dir");
+        }
+
+        if (!$this->core->getAccess()->canI("grading.electronic.grade", ["gradeable" => $gradeable, "graded_gradeable" => $graded_gradeable])) {
+            return JsonResponse::getErrorResponse('You do not have permission to grade this student');
+        }
+
+        $active_version = $graded_gradeable->getAutoGradedGradeable()->getActiveVersion();
+
+        $annotation_gradeable_path = FileUtils::joinPaths($course_path, $target_dir, $gradeable_id);
+        if (!is_dir($annotation_gradeable_path) && !FileUtils::createDir($annotation_gradeable_path)) {
+            return JsonResponse::getErrorResponse('Creating annotation gradeable folder failed');
+        }
+        $annotation_user_path = FileUtils::joinPaths($annotation_gradeable_path, $user_id);
+        if (!is_dir($annotation_user_path) && !FileUtils::createDir($annotation_user_path)) {
+            return JsonResponse::getErrorResponse('Creating annotation user folder failed');
+        }
+        $annotation_version_path = FileUtils::joinPaths($annotation_user_path, $active_version);
+        if (!is_dir($annotation_version_path) && !FileUtils::createDir($annotation_version_path)) {
+            return JsonResponse::getErrorResponse('Creating annotation version folder failed');
+        }
+
+        $decoded_annotations = json_decode($annotations, true) !== null ? json_decode($annotations, true) : [];
+        $annotation_file_path = FileUtils::joinPaths($annotation_version_path, md5($file_path) . "_" . $grader_id . '.json');
+
+        // Check if annotations have no markers (empty annotations)
+        $has_markers = isset($decoded_annotations['markers']) && is_array($decoded_annotations['markers']) && count($decoded_annotations['markers']) > 0;
+
+        if (!$has_markers) {
+            // No markers found - delete the annotation file if it exists
+            if (file_exists($annotation_file_path)) {
+                if (unlink($annotation_file_path)) {
+                    return JsonResponse::getSuccessResponse('Empty annotation removed successfully!');
+                }
+                else {
+                    return JsonResponse::getErrorResponse('Failed to remove empty annotation file');
+                }
+            }
+            // File doesn't exist and there are no markers - nothing to do
+            return JsonResponse::getSuccessResponse('No annotations to save');
+        }
+
+        // Has markers - save the annotation file
+        $annotation_body = [
+            'file_path' => $file_path,
+            'grader_id' => $grader_id,
+            'annotations' => $decoded_annotations
+        ];
+
+        $annotation_json = json_encode($annotation_body);
+
+        if (file_put_contents($annotation_file_path, $annotation_json) === false) {
+            return JsonResponse::getErrorResponse('Failed to save annotation file');
+        }
+
+        return JsonResponse::getSuccessResponse('Image annotation saved successfully!');
+    }
+}
